simplify serialize deserialize
