+++ core/lib.rs
@@ -18,7 +18,6 @@ mod ops;
 pub mod plugin_api;
 mod resources;
 mod runtime;
-mod serialize_deserialize;
 mod shared_queue;
 mod zero_copy_buf;
